Create system.py
